update time_series_helpers with new version
