+++ aurora/pipelines/time_series_helpers.py
@@ -2,18 +2,20 @@ import numpy as np
 import scipy.signal as ssig
 import xarray as xr
 
-from deprecated import deprecated
-
+from aurora.time_series.frequency_domain_helpers import get_fft_harmonics
+from aurora.time_series.windowed_time_series import WindowedTimeSeries
 from aurora.time_series.windowing_scheme import WindowingScheme
 
-@deprecated(version="0.0.3")
 def validate_sample_rate(run_ts, expected_sample_rate):
     """
 
     Parameters
     ----------
-    run_ts
-    config
+    run_ts: mth5.timeseries.run_ts.RunTS
+        Time series object
+    expected_sample_rate: float
+        The samepling rate the time series is expected to have. Normally taken from
+        the processing config
 
     Returns
     -------
@@ -22,65 +24,92 @@ def validate_sample_rate(run_ts, expected_sample_rate):
     if run_ts.sample_rate != expected_sample_rate:
         print(
             f"sample rate in run time series {run_ts.sample_rate} and "
-            f"processing config {expected_sample_rate} do not match"
+            f"processing decimation_obj {decimation_obj.sample_rate} do not match"
         )
         raise Exception
     return
 
 
-@deprecated(version="0.0.3", reason="new mt_metadata based config")
-def apply_prewhitening(config, run_xrts_input):
-    if config["prewhitening_type"] == "first difference":
+def apply_prewhitening(decimation_obj, run_xrts_input):
+    """
+
+    Parameters
+    ----------
+    decimation_obj : aurora.config.metadata.decimation_level.DecimationLevel
+        Information about how the decimation level is to be processed
+    run_xrts_input : xarray.core.dataset.Dataset
+        Time series to be prewhitened
+
+    Returns
+    -------
+    run_xrts : xarray.core.dataset.Dataset
+        prewhitened time series
+
+    """
+    if decimation_obj.prewhitening_type == "first difference":
         run_xrts = run_xrts_input.diff("time")
     else:
+        print(f"{decimation_obj.prewhitening_type} prehitening not yet implemented")
+        print(f"returning original time series")
         run_xrts = run_xrts_input
     return run_xrts
 
 
-@deprecated(version="0.0.3", reason="new mt_metadata based config")
-def apply_recoloring(config, stft_obj):
-    if config["prewhitening_type"] == "first difference":
-        from aurora.time_series.frequency_domain_helpers import get_fft_harmonics
-        from numpy import pi
+def apply_recoloring(decimation_obj, stft_obj):
+    """
+    Parameters
+    ----------
+    decimation_obj : aurora.config.metadata.decimation_level.DecimationLevel
+        Information about how the decimation level is to be processed
+    stft_obj : xarray.core.dataset.Dataset
+        Time series of Fourier coefficients to be recoloured
 
-        freqs = get_fft_harmonics(config.num_samples_window, config.sample_rate)
-        prewhitening_correction = 1.0j * 2 * pi * freqs  # jw
+
+    Returns
+    -------
+    stft_obj : xarray.core.dataset.Dataset
+        Recolored time series of Fourier coefficients
+    """
+    if decimation_obj.prewhitening_type == "first difference":
+        freqs = get_fft_harmonics(decimation_obj.window.num_samples, 
+                                  decimation_obj.decimation.sample_rate)
+        prewhitening_correction = 1.0j * 2 * np.pi * freqs  # jw
         stft_obj /= prewhitening_correction
     return stft_obj
 
 
-@deprecated(version="0.0.3", reason="new mt_metadata based config")
-def run_ts_to_stft_scipy(config, run_xrts_orig):
+def run_ts_to_stft_scipy(decimation_obj, run_xrts_orig):
     """
     Parameters
     ----------
-    config
-    run_xrts
+    decimation_obj : aurora.config.metadata.decimation_level.DecimationLevel
+        Information about how the decimation level is to be processed
+    run_xrts : : xarray.core.dataset.Dataset
+        Time series to be processed
 
     Returns
     -------
-
+    stft_obj : xarray.core.dataset.Dataset
+        Time series of Fourier coefficients
     """
-    import xarray as xr
-
-    run_xrts = apply_prewhitening(config, run_xrts_orig)
+    run_xrts = apply_prewhitening(decimation_obj, run_xrts_orig)
 
     windowing_scheme = WindowingScheme(
-        taper_family=config.taper_family,
-        num_samples_window=config.num_samples_window,
-        num_samples_overlap=config.num_samples_overlap,
-        taper_additional_args=config.taper_additional_args,
-        sample_rate=config.sample_rate,
+        taper_family=decimation_obj.window.type,
+        num_samples_window=decimation_obj.window.num_samples,
+        num_samples_overlap=decimation_obj.window.overlap,
+        taper_additional_args=decimation_obj.window.additional_args,
+        sample_rate=decimation_obj.decimation.sample_rate,
     )
-    # stft_obj = run_xrts.copy(deep=True)
+
     stft_obj = xr.Dataset()
     for channel_id in run_xrts.data_vars:
         ff, tt, specgm = ssig.spectrogram(
             run_xrts[channel_id].data,
-            fs=config.sample_rate,
+            fs=decimation_obj.decimation.sample_rate,
             window=windowing_scheme.taper,
-            nperseg=config.num_samples_window,
-            noverlap=config.num_samples_overlap,
+            nperseg=decimation_obj.window.num_samples,
+            noverlap=decimation_obj.window.overlap,
             detrend="linear",
             scaling="density",
             mode="complex",
@@ -89,14 +118,11 @@ def run_ts_to_stft_scipy(config, run_xrts_orig):
         # drop Nyquist>
         ff = ff[:-1]
         specgm = specgm[:-1, :]
-
-        import numpy as np
-
         specgm *= np.sqrt(2)
 
         # make time_axis
         tt = tt - tt[0]
-        tt *= config.sample_rate
+        tt *= decimation_obj.decimation.sample_rate
         time_axis = run_xrts.time.data[tt.astype(int)]
 
         xrd = xr.DataArray(
@@ -106,19 +132,20 @@ def run_ts_to_stft_scipy(config, run_xrts_orig):
         )
         stft_obj.update({channel_id: xrd})
 
-    stft_obj = apply_recoloring(config, stft_obj)
+    stft_obj = apply_recoloring(decimation_obj, stft_obj)
 
     return stft_obj
 
 
-@deprecated(version="0.0.3", reason="new mt_metadata based config")
-def run_ts_to_stft(config, run_xrts_orig):
+def run_ts_to_stft(decimation_obj, run_xrts_orig):
     """
 
     Parameters
     ----------
-    config : ShortTimeFourierTransformConfig object
-    run_ts ; xarray.core.dataset.Dataset, normally extracted from mth5.RunTS
+    decimation_obj : aurora.config.metadata.decimation_level.DecimationLevel
+        Information about how the decimation level is to be processed
+    run_ts ; xarray.core.dataset.Dataset
+        normally extracted from mth5.RunTS
 
     Returns
     -------
@@ -127,20 +154,22 @@ def run_ts_to_stft(config, run_xrts_orig):
         recoloring. This really doesn't matter since we don't use the DC harmonic for
         anything.
     """
-    from aurora.time_series.windowed_time_series import WindowedTimeSeries
+    try:
+        windowing_scheme = WindowingScheme(
+            taper_family=decimation_obj.window.type,
+            num_samples_window=decimation_obj.window.num_samples,
+            num_samples_overlap=decimation_obj.window.overlap,
+            taper_additional_args=decimation_obj.window.additional_args,
+            sample_rate=decimation_obj.decimation.sample_rate,
+        )
+    except AttributeError:
+        print("AttributeError --- run_ts_to_stft ?")
 
-    windowing_scheme = WindowingScheme(
-        taper_family=config.taper_family,
-        num_samples_window=config.num_samples_window,
-        num_samples_overlap=config.num_samples_overlap,
-        taper_additional_args=config.taper_additional_args,
-        sample_rate=config.sample_rate,
-    )
 
-    run_xrts = apply_prewhitening(config, run_xrts_orig)
+    run_xrts = apply_prewhitening(decimation_obj, run_xrts_orig)
 
     windowed_obj = windowing_scheme.apply_sliding_window(
-        run_xrts, dt=1.0 / config.sample_rate
+        run_xrts, dt=1.0 / decimation_obj.decimation.sample_rate
     )
     windowed_obj = WindowedTimeSeries.detrend(data=windowed_obj, detrend_type="linear")
 
@@ -151,50 +180,33 @@ def run_ts_to_stft(config, run_xrts_orig):
     # scale_factor=windowing_scheme.linear_spectral_density_calibration_factor)
 
     stft_obj = windowing_scheme.apply_fft(
-        tapered_obj, detrend_type=config.extra_pre_fft_detrend_type
+        tapered_obj, detrend_type=decimation_obj.extra_pre_fft_detrend_type
     )
-    stft_obj = apply_recoloring(config, stft_obj)
+    stft_obj = apply_recoloring(decimation_obj, stft_obj)
 
     return stft_obj
 
 
-@deprecated(version="0.0.3", reason="new mt_metadata based config")
-def run_ts_to_calibrated_stft(run_ts, run_obj, config, units="MT"):
-    """
-    Parameters
-    ----------
-    run_ts
-    run_obj
-    config
-    units
-
-    Returns
-    -------
-
-    """
-    stft_obj = run_ts_to_stft(config, run_ts.dataset)
-    stft_obj = calibrate_stft_obj(stft_obj, run_obj, units=units)
-
-    return stft_obj
-
-
-@deprecated(version="0.0.3", reason="new mt_metadata based config")
 def calibrate_stft_obj(stft_obj, run_obj, units="MT", channel_scale_factors=None):
     """
 
     Parameters
     ----------
-    stft_obj
-    run_obj
-    units
-    scale_factors : dict
+    stft_obj : xarray.core.dataset.Dataset
+        Time series of Fourier coefficients to be calibrated
+    run_obj : mth5.groups.master_station_run_channel.RunGroup
+        Provides information about filters for calibration
+    units : string
+        usually "MT", contemplating supporting "SI"
+    scale_factors : dict or None
         keyed by channel, supports a single scalar to apply to that channels data
         Useful for debugging.  Should not be used in production and should throw a
         warning if it is not None
 
     Returns
     -------
-
+    stft_obj : xarray.core.dataset.Dataset
+        Time series of calibrated Fourier coefficients
     """
     for channel_id in stft_obj.keys():
         mth5_channel = run_obj.get_channel(channel_id)
@@ -217,73 +229,64 @@ def calibrate_stft_obj(stft_obj, run_obj, units="MT", channel_scale_factors=None
     return stft_obj
 
 
-@deprecated(version="0.0.3", reason="new mt_metadata based config")
-def get_data_from_mth5(config, mth5_obj, run_id):
+def get_run_run_ts_from_mth5(mth5_obj, station_id, run_id, expected_sample_rate,
+                             start=None, end=None):
     """
-    ToDo: Review if this method should be moved into mth5.  If that were the case,
-    the config being passed here should be replaced with a list of station_ids and
-    the config sampling_rate, so that there is no dependency on the config object in
-    mth5.
-    In a future version this could also take a decimation level as an argument.  It
-    could then be merged with prototype decimate, depending on the decimation level.
+    ToDo: Review if this method should be moved into mth5.
 
-    Parameters
-    ----------
-    config : decimation_level_config
-    mth5_obj
+    Simple implementation of what could eventually become a somewhat complicated
+    function -- see issue #13.  In a future version this could also take a decimation
+    level as an argument.  It could then be merged with prototype decimate, depending
+    on the decimation level.
 
-    Returns
-    -------
-
-    Somewhat complicated function -- see issue #13.  Ultimately this method could be
-    embedded in mth5, where the specific attributes of the config needed for this
-    method are passed as explicit arguments.
-
-    Should be able to
+    Future version should be able to
     1. accept a config and an mth5_obj and return decimation_level_0,
     2. Accept data from a given decimation level, and decimation
     instrucntions and return it
     3. If we decide to house decimated data in an mth5 should return time
     series for the run at the perscribed decimation level
 
-    Thus args are
+    Thus args would be
     decimation_level_config, mth5,
     decimation_level_config, runs and run_ts'
     decimation_level_config, mth5
-    Returns: tuple of dicts
-        Each dictionary is associated with a station, one for local and one
-        for remote at this point
-        Each Dict has keys "run" and "mvts" which are the mth5_run and the
-        mth5_run_ts objects respectively for the associated station
+
+    Parameters
+    ----------
+    mth5_obj: mth5.mth5.MTH5
+        The data container with run and time series'
+    station_id: str
+        The name of the station to get data from
+    run_id: str
+        The name of the run to get data from
+    sample_rate : float (may choose to also support  None)
+        expected sample rate of data in the mth5
+    Returns
     -------
 
+
+
+    Returns
+    -------
+    run_run_ts : dict
+        Dictionary associated with a station-run. Has keys "run" and "mvts".
+        "run" maps to mth5.groups.master_station_run_channel.RunGroup
+        "mvts" maps to xarray.core.dataset.Dataset
+
     """
-    # <LOCAL>
-    local_run_obj = mth5_obj.get_run(config["local_station_id"], run_id)
-    local_run_ts = local_run_obj.to_runts()
-    validate_sample_rate(local_run_ts, config.sample_rate)
-    local = {"run": local_run_obj, "mvts": local_run_ts.dataset, "run_id":run_id}
-    # </LOCAL>
-
-    # <REMOTE>
-    if config.reference_station_id:
-        remote_run_obj = mth5_obj.get_run(config["reference_station_id"], run_id)
-        remote_run_ts = remote_run_obj.to_runts()
-        validate_sample_rate(remote_run_ts, config.sample_rate)
-        remote = {"run": remote_run_obj, "mvts": remote_run_ts.dataset}
-    else:
-        remote = {"run": None, "mvts": None}
-    # </REMOTE>
-    return local, remote
+    run_obj = mth5_obj.get_run(station_id, run_id)
+    run_ts = run_obj.to_runts(start=start, end=end)
+    validate_sample_rate(run_ts, expected_sample_rate)
+    run_run_ts = {"run": run_obj, "mvts": run_ts.dataset}
+    return run_run_ts
 
 
-@deprecated(version="0.0.3", reason="new mt_metadata based config")
 def prototype_decimate(config, run_run_ts):
     """
     TODO: ?Move this function into time_series/decimate.py?
     Parameters
     ----------
-    config : DecimationConfig object
+    config : aurora.config.metadata.decimation.Decimation
     run_run_ts: dict keyed by "run" and "mvts"
     out_dict["run"] is mth5.groups.master_station_run_channel.RunGroup
     out_dict["mvts"] is mth5.timeseries.run_ts.RunTS
@@ -298,7 +301,7 @@ def prototype_decimate(config, run_run_ts):
 
     # <Replace with rolling mean, somethng that works with time>
     # and preferably takes the average time, not the start of the window
-    slicer = slice(None, None, config.decimation_factor)
+    slicer = slice(None, None, int(config.factor))#decimation.factor
     downsampled_time_axis = run_xrts.time.data[slicer]
     # </Replace with rolling mean, somethng that works with time>
 
@@ -307,7 +310,7 @@ def prototype_decimate(config, run_run_ts):
     num_channels = len(channel_labels)
     new_data = np.full((num_observations, num_channels), np.nan)
     for i_ch, ch_label in enumerate(channel_labels):
-        new_data[:, i_ch] = ssig.decimate(run_xrts[ch_label], config.decimation_factor)
+        new_data[:, i_ch] = ssig.decimate(run_xrts[ch_label], int(config.factor))
 
     xr_da = xr.DataArray(
         new_data,
